replaced coinstable with new component
