+++ src/components/CoinsPercentageBar/CoinsPercentageBar.styles.jsx
